fix: Change worklist flags from uint8_t to uint32_t for atomicCAS compatibility

+++ src/gpu/dijkstra/main_gpu.cu
@@ -17,7 +17,7 @@ extern "C" {
         uint32_t* d_new_worklist,
         uint32_t num_current_nodes,
         uint32_t* d_new_worklist_size,
-        uint8_t* d_worklist_flags,
+        uint32_t* d_worklist_flags,
         float delta,
         int block_size);
     
@@ -35,7 +35,7 @@ private:
     uint32_t* d_worklist_1;
     uint32_t* d_worklist_2;
     uint32_t* d_worklist_size;
-    uint8_t* d_worklist_flags;  // Flag array for deduplication
+    uint32_t* d_worklist_flags;  // Flag array for deduplication
     uint32_t max_nodes;
     bool is_initialized;
 
@@ -54,7 +54,7 @@ public:
         CUDA_CHECK(cudaMalloc(&d_worklist_1, max_nodes * sizeof(uint32_t)));
         CUDA_CHECK(cudaMalloc(&d_worklist_2, max_nodes * sizeof(uint32_t)));
         CUDA_CHECK(cudaMalloc(&d_worklist_size, sizeof(uint32_t)));
-        CUDA_CHECK(cudaMalloc(&d_worklist_flags, max_nodes * sizeof(uint8_t)));
+        CUDA_CHECK(cudaMalloc(&d_worklist_flags, max_nodes * sizeof(uint32_t)));
         
         is_initialized = true;
         std::cout << "GPU Dijkstra initialized for " << max_nodes << " nodes" << std::endl;
@@ -113,7 +113,7 @@ public:
             }
             // Reset next worklist size and clear flags for deduplication
             CUDA_CHECK(cudaMemcpy(d_worklist_size, &zero, sizeof(uint32_t), cudaMemcpyHostToDevice));
-            CUDA_CHECK(cudaMemset(d_worklist_flags, 0, max_nodes * sizeof(uint8_t)));
+            CUDA_CHECK(cudaMemset(d_worklist_flags, 0, max_nodes * sizeof(uint32_t)));
             
             // Launch SSSP kernel
             launch_sssp_kernel(
